User admin content

+++ backend/app/api/routes/about/delete.py
@@ -5,8 +5,11 @@ from starlette.status import HTTP_200_OK
 
 from app.api.dependencies.database import get_db_repository
 
+from app.api.dependencies.auth import get_user_from_token, is_superuser, is_verified
+
 from app.db.repositories.about.about import AboutDBRepository
 
+from app.models.user import UserInDB
 
 router = APIRouter()
 
@@ -14,7 +17,14 @@ router = APIRouter()
 async def delete_team_member(
     order: int = Path(...),
     db_repo: AboutDBRepository = Depends(get_db_repository(AboutDBRepository)),
+    user: UserInDB = Depends(get_user_from_token),
+    is_superuser = Depends(is_superuser),
+    is_verified = Depends(is_verified),
     ) -> None:
+    if not user.is_superuser:
+        raise HTTPException(status_code=HTTP_403_FORBIDDEN, detail="Not superuser!")
+    if not is_verified:
+        raise HTTPException(status_code=HTTP_403_FORBIDDEN, detail="Email not verified!")
     
     await db_repo.delete_team_member(id=order)
 
@@ -23,7 +33,14 @@ async def delete_team_member(
 async def delete_contact(
     order: int = Path(...),
     db_repo: AboutDBRepository = Depends(get_db_repository(AboutDBRepository)),
+    user: UserInDB = Depends(get_user_from_token),
+    is_superuser = Depends(is_superuser),
+    is_verified = Depends(is_verified),
     ) -> None:
+    if not user.is_superuser:
+        raise HTTPException(status_code=HTTP_403_FORBIDDEN, detail="Not superuser!")
+    if not is_verified:
+        raise HTTPException(status_code=HTTP_403_FORBIDDEN, detail="Email not verified!")
 
     await db_repo.delete_contact(id=order)
 
@@ -31,7 +48,14 @@ async def delete_contact(
 async def delete_about_project(
     order: int = Path(...),
     db_repo: AboutDBRepository = Depends(get_db_repository(AboutDBRepository)),
+    user: UserInDB = Depends(get_user_from_token),
+    is_superuser = Depends(is_superuser),
+    is_verified = Depends(is_verified),
     ) -> None:
+    if not user.is_superuser:
+        raise HTTPException(status_code=HTTP_403_FORBIDDEN, detail="Not superuser!")
+    if not is_verified:
+        raise HTTPException(status_code=HTTP_403_FORBIDDEN, detail="Email not verified!")
 
     await db_repo.delete_about_project(id=order)
